Consuming Coverage Eligibiity request endpoint when checking eligibility
